+++ packages/edge-config/src/utils/index.ts
@@ -1,19 +1,6 @@
 import type { Connection } from '../types';
 import { trace } from './tracing';
 
-export const ERRORS = {
-  UNAUTHORIZED: '@vercel/edge-config: Unauthorized',
-  EDGE_CONFIG_NOT_FOUND: '@vercel/edge-config: Edge Config not found',
-};
-
-export class UnexpectedNetworkError extends Error {
-  constructor(res: Response) {
-    super(
-      `@vercel/edge-config: Unexpected error due to response with status code ${res.status}`,
-    );
-  }
-}
-
 /**
  * Checks if an object has a property
  */
